+++ package.json
@@ -24,6 +24,7 @@
     "npm": "^11.4.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.2",
     "react-stockcharts": "^0.7.8",
     "tailwind-merge": "^3.3.0",
     "tailwindcss-animate": "^1.0.7",
