+++ src/app/router.tsx
@@ -1,0 +1,16 @@
+import Home from "@/pages/home/Home";
+import Test from "@/pages/test/Test";
+import { createBrowserRouter } from "react-router-dom";
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <Home />,
+  },
+  {
+    path: "/test",
+    element: <Test />,
+  },
+]);
+
+export default router;
