+++ src/pages/home/Home.tsx
@@ -1,0 +1,5 @@
+function Home() {
+  return <p>home</p>;
+}
+
+export default Home;
